fix tests that are broken by jacoco coverage
Jacoco instruments classes and adds in methods to classes under
test. Some tests have faulty assumptions about how many methods are in a
class and simply get the first declared method in a class, which is
sometimes a Jacoco method and sometimes not. This change all known uses
of getMethods()[x] and getDeclaredMethods()[x] to use getMethod() and
getDeclaredMethod to get the desired method.