add support for API keys in Swagger generation
This change doesn't actually add API key validation, since that's done
in the endpoints-management libraries. This simply adds Swagger
generation so that the the management filter knows to do the validation.
To keep things simple, we provide a boolean true or false and fix the
API key to be a query parameter named "key". This is in line with what
client libraries expect.