de-couple new and old audiences in the model
Right now, the legacy audiences directly impact the data structures used
for third party JWT issuers. However, if you add audiences below the
@Api level, then this means some ApiConfig objects will have Google
issuer configs and some will not. This change removes the effect of the
legacy audiences on the new and instead lets other code merge them as
necessary, to keep the model consistent.