fix DiscoveryContext API root computation
This change stops caching the API root in DiscoveryContext and
reconstitutes it when requested. Previously, one could call setHostname,
and it wouldn't update the API root. This change makes it so the API
root from the context is always up to date. This also removes the
dependency on the Swagger Scheme model.