fix discovery schema writing across multiple APIs
Previously, due to how caching was implemented, if a resource was shared
across multiple APIs, only the first API discovery would have the schema
for the fields in the resource. This change adds all relevant schema for
a resource to an API when the resource already exists in the repository.