fix empty servicePath in discovery
The discovery test was not catching this issue in part because it wasn't
using the exact same JSON serialization as it does at runtime. That has
also been fixed. This change will likely reintroduce empty maps in
serialization, but this is currently the best workaround to allow empty
strings inside maps, which we need to serialize the discovery document
correctly.