add prettyPrint servlet initialization parameter
This new parameter is a boolean value that controls whether or not
prettyPrint is enabled by default. Requests can still enable pretty
printing by adding a query parameter if necessary. This value is set to
true by default.

This value is propagated for each request via EndpointsContext, which
requires some amount of refactoring to plumb the value to
StandardParameters.shouldPrettyPrint.