update OpenAPI spec generation to new auth format
The new auth format specifies audiences in the OpenAPI security
definitions block, rather than using the x-security vendor
extension. To handle this, we now declare one security definition per
issuer-audience set combination. The security definition is given the
name of the issuer plus a hash of the audience set in hexadecimal.