quietly allow underscores in API names
The official documentation for a long time has disallowed underscores
in API names, so I'm leaving it as such. However, for extremely old
users of the old frameworks, this limitation was not in
place. Therefore, I'm removing it from the API validation, but leaving
the @Api documentation in order to not encourage the style.