add option to disable Jackson annotations
Endpoints previously bundled the Jackson annotation introspector,
which didn't really have an effect when Jackson was a vendored and
repackaged dependency. Now that it's an explicit dependency, the
annotations can clash with previous Endpoints behavior when the
annotations are used outside of an Endpoints context.

This change adds an option to disable the Jackson behavior, which is
done so by setting ENDPOINTS_JSON_DISABLE_JACKSON_ANNOTATIONS to a
truthy value or the endpoints.json.disableJacksonAnnotations system
property.

Consequently, this change also refactors MapSchemaFlag into a general
EndpointsFlag class for future use, as it is more flexible and easy to
implement than the ServletInitializationParameters flags.